Исправил опечатку в типе функции

+++ laboratory-task-7/worckWithLines.hpp
@@ -5,7 +5,7 @@
 void getText(char* text,const char* textToWrite);
 bool isPalindrom(char* word);
 void printMaxPolindroms(char* text, char* delimiters, char* newLine);
-int seekMaxLenPolindrom(char* text, char* delimiters);
+size_t seekMaxLenPolindrom(char* text, char* delimiters);
 bool checkAllSpace(char* text);
 
 #endif
